[Test] Consolidate cross entropy tests
